backend/requirements.txt
fastapi[standard]==0.115.0
pydantic-settings==2.5.2
alembic==1.13.3
sqlmodel==0.0.22
tenacity==9.0.0
requests==2.32.3
requests_toolbelt==1.0.0
pygithub==2.4.0
gql==3.5.0
